Moved trainee management to separate blueprint
