+++ collectives/__init__.py
@@ -19,7 +19,8 @@ from flask_migrate import Migrate
 from flask_statistics import Statistics
 
 from . import models, api, forms
-from .routes import root, profile, auth, administration, event, payment, technician
+from .routes import root, profile, auth, administration, event, payment, technician, activity_supervison
+from .routes import activity_supervison
 from .utils import extranet, init, jinja, error, access, payline, statistics
 
 
@@ -85,6 +86,7 @@ def create_app(config_filename="config"):
         app.register_blueprint(event.blueprint)
         app.register_blueprint(payment.blueprint)
         app.register_blueprint(technician.blueprint)
+        app.register_blueprint(activity_supervison.blueprint)
 
         # Error handling
         app.register_error_handler(werkzeug.exceptions.NotFound, error.not_found)
