+++ collectives/templates/trainees.html
@@ -24,7 +24,6 @@
 
 {% block content %}
 <div class="panel" id="administration">
-  <h3>Administration</h3>
   <h4>Initiateurs en formation</h4>
 
   <div class="action">
@@ -41,7 +40,7 @@
             {% endfor %}
         </div>
 
-      <form action="{{url_for('administration.add_trainee')}}" id="user-search-form" method="POST" >
+      <form action="{{url_for('activity_supervision.add_trainee')}}" id="user-search-form" method="POST" >
         <div id="user-search-data" class="form-search search-data">
         {{ add_trainee_form.activity_id }}
         <input id = "user-search"  autocomplete="off" class="search-input" placeholder="Nom..." >
